Fix discrepancy in refactoring NotificationUtil
